add file, tasklist, and tasks

+++ src/db/tasks.ts
@@ -4,8 +4,8 @@ import { getUuid } from "../util"
 
 export interface BasicTaskObject {
   description: string
-  due: string | null
-  completed: string | null
+  due: Date | null
+  completed: Date | null
 }
 
 export interface TaskObject extends BasicTaskObject {
@@ -14,15 +14,28 @@ export interface TaskObject extends BasicTaskObject {
 }
 
 export interface TaskController {
+  checkAccess (accountId: string, taskId: string): Promise<{ access: boolean, exists: boolean }>
   createTask (taskListId: string, task: BasicTaskObject): Promise<TaskObject>
   deleteTask (taskId: string): Promise<void>
   deleteTasks (conn: PoolClient, taskListId: string): Promise<void>
   getTask (taskId: string): Promise<TaskObject>
   getTasks (taskListId: string): Promise<Array<TaskObject>>
-  setTask (task: TaskObject): Promise<TaskObject>
+  setTask (taskId: string, task: BasicTaskObject): Promise<TaskObject>
 }
 
 export function TaskFactory (db: Pool, controller: DatabaseController): TaskController {
+
+  async function checkAccess (accountId: string, taskId: string): Promise<{ access: boolean, exists: boolean }> {
+    const { rows } = await db.query({
+      name: 'has-task-access',
+      text: 'SELECT L.account_id as "account_id" FROM tasks T INNER JOIN task_lists L ON T.task_list_id = L.id WHERE T.id = $1',
+      values: [taskId]
+    })
+
+    return rows.length > 0
+      ? { exists: true, access: rows[0].account_id === accountId }
+      : { exists: false, access: false }
+  }
 
   async function createTask (taskListId: string, task: BasicTaskObject): Promise<TaskObject> {
     const id = getUuid()
@@ -42,29 +55,50 @@ export function TaskFactory (db: Pool, controller: DatabaseController): TaskCont
   }
 
   async function deleteTask (taskId: string): Promise<void> {
-    await db.query({
-      name: 'delete-tasks',
-      text: 'DELETE FROM tasks WHERE id = $1',
-      values: [taskId]
+    return controller.transaction(null, async conn => {
+      await db.query({
+        name: 'delete-tasks',
+        text: 'DELETE FROM tasks WHERE id = $1',
+        values: [taskId]
+      })
+  
+      await controller.files.deleteFiles(conn, taskId)
     })
-
-    // TODO: delete files
   }
 
   async function deleteTasks (conn: PoolClient, taskListId: string): Promise<void> {
-    await conn.query({
-      name: 'delete-tasklist-tasks',
-      text: 'DELETE FROM tasks WHERE task_list_id = $1',
-      values: [taskListId]
+    return controller.transaction(conn, async conn => {
+      const { rows } = await conn.query({
+        name: 'delete-tasklist-tasks',
+        text: 'DELETE FROM tasks WHERE task_list_id = $1 RETURNING *',
+        values: [taskListId]
+      })
+  
+      const length = rows.length
+      if (length) {
+        for (let i = 0; i < length; i++) {
+          await controller.files.deleteFiles(conn, rows[0].id)
+        }
+      }
+      
     })
-
-    // TODO: delete files
   }
 
   async function getTask (taskId: string): Promise<TaskObject> {
     const { rows } = await db.query({
       name: 'get-task',
-      text: 'SELECT * FROM tasks as T, files as F WHERE T.id = $1 AND T.id = F.task_id',
+      text: `SELECT 
+          T.id as "task_id",
+          task_list_id, 
+          description, 
+          due, 
+          completed, 
+          F.id as "file_id", 
+          F.name as "file_name", 
+          file_path 
+        FROM tasks T 
+        LEFT JOIN files F ON T.id = F.task_id
+        WHERE T.id = $1`,
       values: [taskId]
     })
     return formatTasks(rows)[0] || null
@@ -73,22 +107,34 @@ export function TaskFactory (db: Pool, controller: DatabaseController): TaskCont
   async function getTasks (taskListId: string): Promise<Array<TaskObject>> {
     const { rows } = await db.query({
       name: 'get-tasks',
-      text: 'SELECT * FROM tasks as T, files as F WHERE T.task_list_id = $1 AND T.id = F.task_id',
+      text: `SELECT 
+          T.id as "task_id",
+          task_list_id, 
+          description, 
+          due, 
+          completed, 
+          F.id as "file_id", 
+          F.name as "file_name", 
+          file_path 
+        FROM tasks T 
+        LEFT JOIN files F ON T.id = F.task_id
+        WHERE T.task_list_id = $1`,
       values: [taskListId]
     })
     return formatTasks(rows)
   }
 
-  async function setTask (task: TaskObject): Promise<TaskObject> {
-    await db.query({
+  async function setTask (taskId: string, task: BasicTaskObject): Promise<TaskObject> {
+    const { rows } = await db.query({
       name: 'set-task',
-      text: 'UPDATE tasks SET description = $1, due = $2, completed = $3 WHERE id = $4',
-      values: [task.description, task.due, task.completed, task.id]
+      text: 'UPDATE tasks SET description = $1, due = $2, completed = $3 WHERE id = $4 RETURNING *',
+      values: [task.description, task.due, task.completed, taskId]
     })
-    return getTask(task.id)
+    return formatTasks(rows)[0] || null
   }
 
   return {
+    checkAccess,
     createTask,
     deleteTask,
     deleteTasks,
@@ -103,22 +149,32 @@ function formatTasks (rows: Array<QueryResultRow>): Array<TaskObject> {
   const results: Array<TaskObject> = []
   
   rows.forEach(row => {
-    const id = row.T.id
-    const file = {
-      id: row.F.id,
-      name: row.F.name
+    const taskId = row.task_id
+    const fileId = row.file_id
+
+    const task: TaskObject = {
+      id: row.task_id,
+      description: row.description,
+      due: row.due,
+      completed: row.completed,
+      files: []
     }
-    if (filesMap[id]) {
-      filesMap[id].push(file)
-    } else {
-      filesMap[id] = [file]
-      results.push({
-        id: row.T.id,
-        description: row.T.description,
-        due: row.T.due,
-        completed: row.T.completed,
-        files: filesMap[id]
-      })
+
+    // if there is a file then store it
+    if (fileId) {
+      const file = {
+        id: row.file_id,
+        name: row.file_name
+      }
+      if (filesMap[taskId]) {
+        filesMap[taskId].push(file)
+      } else {
+        filesMap[taskId] = [file]
+        task.files = filesMap[taskId]
+        results.push(task)
+      }
+    } else{
+      results.push(task)
     }
   })
